* requests.c (DECLARE): Fix prototype for dispatch functions to 	properly match what is expected by ss library.
	* pager.c (ss_pager_create): Cast return from fork to int for use
	in switch statement.

	* ss.h: Provide full prototype for ss_name(). Add prototypes for
	ss_execute_line(), ss_delete_invocation(), ss_add_info_dir(),
	ss_delete_info_dir(), ss_abort_subsystem(), ss_get_prompt(),
	ss_set_prompt(), ss_add_request_table(),
	ss_delete_request_table(), ss_quit().

	* ss_internal.h: Made full prototypes for ss_parse(),
	ss_page_stdin(), ss_pagre_create(). Add prototypes for
	ss_self_identify(), ss_subsystem_name(), ss_subsystem_version(),
	ss_unimplemented(). Moved ss_execute_line() to ss.h

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12598 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/ss/pager.c
@@ -36,7 +36,7 @@ int ss_pager_create()
 	if (pipe(filedes) != 0)
 		return(-1);
 
-	switch(fork()) {
+	switch((int) fork()) {
 	case -1:
 		return(-1);
 	case 0:
@@ -102,7 +102,7 @@ void ss_page_stdin()
 		char buf[80];
 		register int n;
 		while ((n = read(0, buf, 80)) > 0)
-			write(1, buf, n);
+			write(1, buf, (unsigned) n);
 	}
 	exit(errno);
 }
